[WIP] Initial Commit
The basic structure of the project using webpack for building.

Uses prettier for linting, but maybe it should also use eslint?  I will have to
open this up in VSCode to see how it works there.

testing is done with `tap` to ensure that the project loads in node and puppeteer
will probably be used for integration testing.

I spent a lot of time playing around with both the encryption and the hypercores
and withng out how to write protobufs to the cores.

This starts out by creating a Terminal, or an encrypted area that houses
Identities and hosts hypercores by discovery key, an Identity tells the core to
cache a discoveryKey and the Terminal caches it, meanwhile the Identity can
unlock data to present a social network state.

I need to clean up the code and abstract secure core creation then I can begin
working on the Identity portions